Refactor to improve readability
Problem:
When reading the code, you would notice that there were
far too many nested if..statements at the point where
validaton is finished and the file is about to be moved
to the upload destination. I'm referring to somewhere
around line 64/65. There was about 3 level deep if..statements.

Solution:
Now when the validation is completed, there is only one main
if..statement with the rest of the upload login nested in it.
This is rather simple and easy on the eye.